// src/models/IoK8sApiStorageV1alpha1VolumeAttributesClass/types.ts
import { IoK8sApimachineryPkgApisMetaV1ObjectMetaModelOptions as IoK8sApimachineryPkgApisMetaV1ObjectMeta } from '../IoK8sApimachineryPkgApisMetaV1ObjectMeta';

export interface IoK8sApiStorageV1alpha1VolumeAttributesClassModelOptions {
  apiVersion?: string;
  driverName: string;
  kind?: string;
  metadata?: IoK8sApimachineryPkgApisMetaV1ObjectMeta;
  parameters?: { [key: string]: string };
}