// src/models/IoK8sApiStorageV1alpha1VolumeAttributesClassList/types.ts
import { IoK8sApiStorageV1alpha1VolumeAttributesClassModelOptions as IoK8sApiStorageV1alpha1VolumeAttributesClass } from '../IoK8sApiStorageV1alpha1VolumeAttributesClass';
import { IoK8sApimachineryPkgApisMetaV1ListMetaModelOptions as IoK8sApimachineryPkgApisMetaV1ListMeta } from '../IoK8sApimachineryPkgApisMetaV1ListMeta';

export interface IoK8sApiStorageV1alpha1VolumeAttributesClassListModelOptions {
  apiVersion?: string;
  items: Array<IoK8sApiStorageV1alpha1VolumeAttributesClass>;
  kind?: string;
  metadata?: IoK8sApimachineryPkgApisMetaV1ListMeta;
}